request

+++ mars/comm/jni/OnAlarm.inl
@@ -10,6 +10,6 @@
 
 extern "C" JNIEXPORT void JNICALL Java_com_tencent_mars_comm_Alarm_onAlarm(JNIEnv *, jclass, jlong id)
 {
-    xdebug2(TSF"BroadcastMessage seq:%_", (long long)id);
-    MessageQueue::BroadcastMessage(MessageQueue::GetDefMessageQueue(), MessageQueue::Message(KALARM_MESSAGETITLE, (long long)id, 0));
+    xdebug2(TSF"BroadcastMessage seq:%_", (int64_t)id);
+    MessageQueue::BroadcastMessage(MessageQueue::GetDefMessageQueue(), MessageQueue::Message(KALARM_MESSAGETITLE, (int64_t)id, 0));
 }
